requirements.txt
requests==2.29.0
beautifulsoup4
streamlit
numpy
Pillow
flask
tensorflow==2.6.0
keras==2.6.0